fix: make createCalendarAppointment an internal function, not a public action
